Move core template tag test to core test suite
